alle userstories implementiert, javadoc fehlt noch

+++ src/Listbuilder/Business/Listbuilder.java
@@ -2,7 +2,6 @@ package Listbuilder.Business;
 
 import Listbuilder.Api.InputReader;
 import Listbuilder.Api.MenuPrinter;
-import Listbuilder.Model.Eintrag;
 import Listbuilder.Model.Liste;
 import Listbuilder.Repository.EintragRepository;
 import Listbuilder.Repository.ListeRepository;
@@ -23,8 +22,8 @@ public class Listbuilder {
     public Listbuilder() {
         this.printer = new MenuPrinter();
         this.reader = new InputReader();
-        this.listeRepository = new ListeRepository();
         this.eintragRepository = new EintragRepository();
+        this.listeRepository = new ListeRepository(this.eintragRepository);
     }
 
     public void start() {
@@ -49,8 +48,14 @@ public class Listbuilder {
                 break;
             case 3:
                 printer.printAbout();
+                showMainOptions();
+                break;
+            case 4:
+                System.exit(0);
                 break;
             default:
+                printer.printInvalidInput();
+                showMainOptions();
                 break;
         }
     }
@@ -59,28 +64,31 @@ public class Listbuilder {
         printer.printEnterListeName();
         String listeName = reader.readString();
         listeRepository.createListe(listeName);
+        printer.printListeWasCreated();
         showMainOptions();
     }
 
     private void showAllListen() {
         List<Liste> listen = listeRepository.getAllListen();
-        for (int i = 0; i < listen.size(); i++) {
-            printer.printListe(i + 1, listen.get(i));
-        }
+        checkHasListe(listen);
+        printer.printAllListen(listen);
         printer.printWhichListeToOpen();
         int listeToOpenIndex = reader.readInteger();
-        Liste liste = listeRepository.getAllListen().get(listeToOpenIndex - 1);
-        showListeDetails(liste);
+        openListe(listeToOpenIndex);
+    }
+
+    private void openListe(int index) {
+        try {
+            Liste liste = listeRepository.getAllListen().get(index - 1);
+            showListeDetails(liste);
+        } catch (IndexOutOfBoundsException e) {
+            printer.printInvalidInput();
+            showAllListen();
+        }
     }
 
     private void showListeDetails(Liste liste) {
-        printer.printListeName(liste);
-        List<Eintrag> eintraegeinListe = liste.getEintraege();
-        if (eintraegeinListe.size() > 0) {
-            for (int i = 0; i < eintraegeinListe.size(); i++) {
-                printer.printEintrag(i + 1, eintraegeinListe.get(i));
-            }
-        }
+        printer.printListeDetails(liste);
         showListeOptions(liste.getId());
     }
 
@@ -100,12 +108,20 @@ public class Listbuilder {
                 setEintragErledigt(listeId);
                 break;
             case 3:
-                setListeErledigt(listeId);
+                deleteEintrag(listeId);
                 break;
             case 4:
+                setListeErledigt(listeId);
+                break;
+            case 5:
                 deleteListe(listeId);
                 break;
+            case 6:
+                showMainOptions();
+                break;
             default:
+                printer.printInvalidInput();
+                showListeOptions(listeId);
                 break;
         }
     }
@@ -118,26 +134,69 @@ public class Listbuilder {
     }
 
     private void setEintragErledigt(UUID listeId) {
-        printer.printWhichEintragToMark();
-        int eintragToMark = reader.readInteger();
         Liste listeToGetEintragMarked = listeRepository.getListe(listeId);
-        UUID eintragId = listeToGetEintragMarked.getEintraege().get(eintragToMark - 1).getId();
-        eintragRepository.markEintragAsErledigt(eintragId);
+        checkHasEintrag(listeToGetEintragMarked);
+        printer.printWhichEintragToMark();
+        int eintragToMarkIndex = reader.readInteger();
+        UUID eintragId = getEintragIdFromListeByIndex(listeToGetEintragMarked, eintragToMarkIndex);
+        eintragRepository.toggleEintragErledigt(eintragId);
+        printer.printEintragWasAdded();
         showListeDetails(listeToGetEintragMarked);
     }
 
+    private UUID getEintragIdFromListeByIndex(Liste liste, int index) {
+        try {
+            return liste.getEintraege().get(index - 1).getId();
+        } catch (IndexOutOfBoundsException e) {
+            printer.printInvalidInput();
+            showListeDetails(liste);
+        }
+        return null;
+    }
+
+    private void deleteEintrag(UUID listeId) {
+        Liste listeToGetEintragDeleted = listeRepository.getListe(listeId);
+        checkHasEintrag(listeToGetEintragDeleted);
+        printer.printWhichEintragToDelete();
+        int eintragToDeleteIndex = reader.readInteger();
+        UUID eintragId = getEintragIdFromListeByIndex(listeToGetEintragDeleted, eintragToDeleteIndex);
+        printer.printConfirmAction();
+        if (reader.confirmInput()) {
+            listeRepository.deleteEintragFromListe(listeId, eintragId);
+            printer.printEintragWasDeleted();
+            showListeDetails(listeToGetEintragDeleted);
+        } else {
+            showListeOptions(listeId);
+        }
+    }
+
     private void setListeErledigt(UUID listeId) {
-        listeRepository.markListeAsErledigt(listeId);
+        listeRepository.toggleListeErledigt(listeId);
         showMainOptions();
     }
 
     private void deleteListe(UUID listeId) {
-        listeRepository.deleteListe(listeId);
-        showMainOptions();
+        printer.printConfirmAction();
+        if (reader.confirmInput()) {
+            listeRepository.deleteListe(listeId);
+            printer.printListeWasDeleted();
+            showMainOptions();
+        } else {
+            showListeOptions(listeId);
+        }
     }
 
-    public void clearScreen() {
-        System.out.print("\033[H\033[2J");
-        System.out.flush();
+    private void checkHasListe(List<Liste> listen) {
+        if (listen.size() <= 0) {
+            printer.printNoListeExists();
+            showMainOptions();
+        }
+    }
+
+    private void checkHasEintrag(Liste liste) {
+        if (liste.getEintraege().size() <= 0) {
+            printer.printNoEintragExists();
+            showListeDetails(liste);
+        }
     }
 }
